# Enable more Rust optimizations