--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,3 +1,7 @@
 [workspace]
 members = ["crates/*"]
 resolver = "2"
+
+[profile.release]
+codegen-units = 1
+lto = true