--- a/packages/gradbench/src/App.tsx
+++ b/packages/gradbench/src/App.tsx
@@ -189,10 +189,13 @@ const App = () => {
               Dex
             </a>
             . The Adroit compiler is <span className="bold">fast</span>: it can
-            lex, parse, and typecheck over 450k lines of code per second using
-            one core on a modern laptop, causing little overhead as an initial
-            preprocessing step before passing typed IR as JSON to another tool
-            to perform automatic differentiation.
+            lex, parse, and typecheck{" "}
+            <span className="bold">
+              half a million lines of code per second
+            </span>{" "}
+            using one core on a modern laptop, causing little overhead as an
+            initial preprocessing step before passing typed IR as JSON to
+            another tool to perform automatic differentiation.
           </div>
         </div>
       </div>